+++ tests/main_test.c
@@ -1,5 +1,6 @@
 #include "CuTest.h"
 #include "file_util_test.h"
+#include "test_assistance_test.h"
 #include <stdio.h>
 
 /*
@@ -18,7 +19,8 @@ void RunAllTests(void) {
     CuString *output = CuStringNew();
     CuSuite * suite = CuSuiteNew();
 
-    CuSuiteAddSuite(suite, fileUtilGetSuite());
+    /*CuSuiteAddSuite(suite, fileUtilGetSuite());*/
+    CuSuiteAddSuite(suite, testAssistanceTestsGetSuite());
 
     CuSuiteRun(suite);
     CuSuiteSummary(suite, output);